Rename CommentDisplay to CommentList
